+++ dddplus-visualization/src/main/java/io/github/dddplus/ast/model/KeyRelationEntry.java
@@ -18,6 +18,7 @@ public class KeyRelationEntry {
     private KeyRelation.Type type;
     private String remark;
     private String leftClassPackageName;
+    private String rightClassPackageName;
     private String javadoc;
     private boolean contextual = false;